refactor(settings): extract table actions and simplify test seams
- Extract OverridesTableActions from OverridesGroupBuilder (~200 lines)
- Replace MouseEvent-based clickLabelForTest with triggerActionForTest seam
- Update test reflection to access tableActions field

Impact: OverridesGroupBuilder drops from ~1138 to ~950 lines, test seams follow established VcsColorPanel pattern
